+++ packages/ui/src/ui/context-menu.tsx
@@ -8,6 +8,7 @@ import {
   IconCircleFilled,
 } from "@tabler/icons-react";
 import { cn } from "../utils/cn";
+import { menuContentClass } from "./_menu-classes";
 
 const ContextMenu = ContextMenuPrimitive.Root;
 const ContextMenuTrigger = ContextMenuPrimitive.Trigger;
@@ -43,10 +44,7 @@ const ContextMenuSubContent = React.forwardRef<
 >(({ className, ...props }, ref) => (
   <ContextMenuPrimitive.SubContent
     ref={ref}
-    className={cn(
-      "glass-panel-strong z-50 w-40 overflow-hidden rounded-lg p-1.5 text-overlay-foreground outline-none data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2",
-      className,
-    )}
+    className={cn(menuContentClass, "w-40", className)}
     {...props}
   />
 ));
@@ -59,10 +57,7 @@ const ContextMenuContent = React.forwardRef<
   <ContextMenuPrimitive.Portal>
     <ContextMenuPrimitive.Content
       ref={ref}
-      className={cn(
-        "glass-panel-strong z-50 w-40 overflow-hidden rounded-lg p-1.5 text-overlay-foreground outline-none data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2",
-        className,
-      )}
+      className={cn(menuContentClass, "w-40", className)}
       {...props}
     />
   </ContextMenuPrimitive.Portal>